updta

+++ hx/lualib/json/values.hpp
@@ -164,7 +164,7 @@ namespace values {
 			return true;
 		}
 		bool EndArray(rapidjson::SizeType elementCount) {
-			assert(elementCount == context_.index_);
+			assert((int)elementCount == context_.index_);
 			context_ = stack_.back();
 			stack_.pop_back();
 			context_.submit(L);
